style : window style 추가
