--- apps/web/app/contents/page.tsx
@@ -1,5 +1,0 @@
-import ContentList from '@/containers/ContentList';
-
-export default function Page() {
-  return <ContentList />;
-}
