+++ apps/web/components/IconButtonBox/index.module.scss
@@ -29,7 +29,7 @@ $icon-size: 1.2rem;
     background-color: var(--color-gray-9);
   }
   .isOn {
-    background-color: var(--color-primary);
+    background-color: var(--color-bg-primary);
   }
 }
