+++ apps/web/components/Items/index.module.scss
@@ -4,9 +4,12 @@ $icon-size: 0.8rem;
 
 .wrap {
   position: relative;
+  width: 100%;
 
   .imgWrap {
+    width: 100%;
     height: 220px;
+    min-width: 250px;
 
     position: relative;
     overflow: hidden;
