+++ apps/web/components/WindowStyles/index.module.scss
@@ -1,0 +1,35 @@
+@use 'styles/mixins' as m;
+
+.wrap {
+  @include m.borderStyles();
+  box-shadow: 3px 4px var(--border-color);
+
+  .titleBar {
+    border-bottom: 2px solid var(--border-color);
+    background-color: var(--color-bg-yellow);
+    padding: 2px 10px;
+
+    @include m.flexbox(between, end);
+
+    h3 {
+      font-size: var(--font-size-m);
+    }
+
+    .barBtn {
+      @include m.flexbox(end);
+      gap: 6px;
+
+      span {
+        height: 24px;
+        font-size: var(--font-size-s);
+        font-weight: var(--font-weight-bold);
+        color: var(--font-color);
+      }
+    }
+  }
+  .clientArea {
+    @include m.flexboxWrap($gap: 20px);
+    padding: 10px;
+    width: 100%;
+  }
+}
