--- apps/web/containers/ContentList/index.module.scss
@@ -1,13 +1,0 @@
-@use 'styles/mixins' as m;
-
-.wrap {
-  width: 100%;
-
-  @include m.flexboxColumn(_, between);
-  gap: 20px;
-
-  .btnBox {
-    width: 100%;
-    @include m.flexbox(around);
-  }
-}
