+++ apps/web/containers/Home/index.module.scss
@@ -1,7 +1,21 @@
 @use 'styles/mixins' as m;
 
-.bannerWrap {
-  width: 100%;
-  height: 200px;
-  background-color: antiquewhite;
+.wrap {
+  @include m.flexboxColumn(_, between);
+  gap: 20px;
+
+  .bannerWrap {
+    width: 100%;
+    height: 200px;
+    background-color: antiquewhite;
+  }
+
+  .btnBox {
+    width: 100%;
+    @include m.flexbox(around);
+  }
+  .clientArea {
+    @include m.flexbox();
+    width: 100%;
+  }
 }
