+++ apps/web/containers/Mypage/index.tsx
@@ -12,7 +12,7 @@ import { useEffect, useState } from 'react';
 import styles from './index.module.scss';
 import ResultList from './ResultList';
 import Loading from '@/components/Loading';
-import { FloatTopBtn } from '@/components/Buttons/FloatTopBtn';
+import FloatTopBtn from '@/components/Buttons/FloatTopBtn';
 import RandomButton from '@/components/Buttons/RandomBtn';
 
 interface UserProps {
