+++ apps/web/containers/layout/Footer/index.module.scss
@@ -5,12 +5,12 @@ $icon-size: 1rem;
 .wrap {
   width: 100%;
   padding: 10px;
-  background-color: var(--color-primary);
+  background-color: var(--color-bg-pink);
   @include m.flexboxColumn(between, end);
 
   p {
     font-size: var(--font-size-s);
-    color: var(--color-white);
+    color: var(--color-font);
   }
 
   .contentsBox {
@@ -39,7 +39,7 @@ $icon-size: 1rem;
       gap: 8px;
 
       span {
-        color: var(--color-white);
+        color: var(--color-font);
         font-size: var(--font-size-xs);
       }
     }
