+++ apps/web/containers/layout/Header/index.module.scss
@@ -7,7 +7,7 @@ $title-width: 220px;
   width: 100%;
   height: 45px;
   padding: 10px 15px;
-  background-color: var(--color-bg-primary);
+  background-color: var(--color-bg-pink);
 
   .titleBtn {
     width: $title-width;
@@ -22,6 +22,7 @@ $title-width: 220px;
 
   .headerBox {
     @include m.flexbox(between);
+
     .btnBox {
       @include m.flexbox(between);
       gap: 20px;
