+++ apps/web/styles/_globals.scss
@@ -13,7 +13,7 @@
   .content_wrap {
     padding: 10px;
     min-height: 640px;
-    margin : 40px 0 ;
+    margin: 60px 0 50px;
   }
 }
